test(balancing): isolate execution language contracts
