+++ libs/gda-balancing/tests/test_http_service.py
@@ -17,49 +17,15 @@ from urllib.error import HTTPError
 from urllib.request import Request, urlopen
 
 import pytest
-from pydantic import ValidationError
 from starlette.types import Receive, Scope, Send
 
 from gda_balancing.application.execution_sessions import (
     ExecutionSessionCreated,
-    ExecutionSessionNotFound,
     ExecutionSessions,
     ExperimentRevisionAdmitted,
-    ExperimentRevisionNotFound,
-)
-from gda_balancing.application.experiment_execution import (
-    ExperimentExecutionRefusal,
-    ExperimentExecutionSuccess,
-    ExperimentExecutionVerdict,
 )
 from gda_balancing.domain.authority.context import packaged_authority_context
-from gda_balancing.domain.diagnostics import (
-    ArtifactLocation,
-    Schema2Diagnostic,
-    Schema2RefusalReport,
-)
-from gda_balancing.domain.publication_types import PublicationMember
 from gda_balancing.domain.runtime.projections import evaluator_build_identity
-from gda_balancing.interfaces.execution_service_language import (
-    EXECUTION_SERVICE_LANGUAGE_REVISION,
-    AdmitExperimentRevisionRequest,
-    CreateExecutionSessionRequest,
-    ExecutionServiceErrorCode,
-    ExecutionSessionCreatedResponse,
-    ExecutionSessionDeletedResponse,
-    ExperimentRevisionAdmittedResponse,
-    ReleaseExecutionSessionRequest,
-    RefusalResponse,
-    RunExperimentRequest,
-    RunRefusalResponse,
-    RunSuccessResponse,
-    RunVerdictResponse,
-    admit_experiment_revision_response,
-    establish_session_response,
-    execution_service_error,
-    execution_service_error_from_condition,
-    run_experiment_revision_response,
-)
 from gda_balancing.interfaces.http.api_v1 import create_api_v1
 from gda_balancing.interfaces.http.local_host import (
     LocalHostReadiness,
@@ -108,272 +74,6 @@ def _rpg_combat_documents() -> tuple[dict[str, Any], dict[str, Any]]:
             (_RPG_COMBAT_EXAMPLE / "experiment.json").read_text(encoding="utf-8")
         ),
     )
-
-
-def _example_refusal() -> Schema2RefusalReport:
-    return Schema2RefusalReport(
-        stage="ingress",
-        diagnostics=(
-            Schema2Diagnostic(
-                code="language.invalid_source",
-                message="the source is invalid",
-                primary=ArtifactLocation(
-                    content_identity="sha256:source",
-                    pointer="/",
-                ),
-            ),
-        ),
-        truncated=False,
-    )
-
-
-def test_execution_service_language_establish_session_contract_is_closed() -> None:
-    payload = {
-        "model_source": {"schema_version": "2.0.0"},
-        "experiment_specification": {"schema_version": "2.0.0"},
-    }
-
-    request = CreateExecutionSessionRequest.model_validate(payload)
-    response = ExecutionSessionCreatedResponse(
-        session_id="session-1",
-        resolved_model_identity="sha256:resolved-model",
-        revision_id="sha256:experiment",
-    )
-
-    assert EXECUTION_SERVICE_LANGUAGE_REVISION == 1
-    assert request.model_dump(mode="json") == payload
-    assert response.model_dump(mode="json") == {
-        "outcome": "success",
-        "session_id": "session-1",
-        "resolved_model_identity": "sha256:resolved-model",
-        "revision_id": "sha256:experiment",
-    }
-    with pytest.raises(ValidationError):
-        CreateExecutionSessionRequest.model_validate(
-            {**payload, "implicit_active_revision": True}
-        )
-
-
-def test_execution_service_language_frames_establish_session_results() -> None:
-    created = ExecutionSessionCreated(
-        session_id="session-1",
-        resolved_model_identity="sha256:resolved-model",
-        revision_id="sha256:experiment",
-    )
-    refusal = _example_refusal()
-
-    success = establish_session_response(created)
-    refused = establish_session_response(refusal)
-
-    assert success == ExecutionSessionCreatedResponse(
-        session_id="session-1",
-        resolved_model_identity="sha256:resolved-model",
-        revision_id="sha256:experiment",
-    )
-    assert refused == RefusalResponse(refusal=refusal)
-
-
-def test_execution_service_language_admit_revision_contract_is_closed() -> None:
-    payload = {
-        "session_id": "session-1",
-        "experiment_specification": {"schema_version": "2.0.0"},
-    }
-
-    request = AdmitExperimentRevisionRequest.model_validate(payload)
-    response = ExperimentRevisionAdmittedResponse(
-        revision_id="sha256:experiment",
-        created=True,
-    )
-
-    assert request.model_dump(mode="json") == payload
-    assert response.model_dump(mode="json") == {
-        "outcome": "success",
-        "revision_id": "sha256:experiment",
-        "created": True,
-    }
-    with pytest.raises(ValidationError):
-        AdmitExperimentRevisionRequest.model_validate(
-            {**payload, "replace_active_revision": True}
-        )
-
-
-def test_execution_service_language_frames_revision_admission_results() -> None:
-    admitted = ExperimentRevisionAdmitted(
-        revision_id="sha256:experiment",
-        created=True,
-    )
-    refusal = _example_refusal()
-
-    success = admit_experiment_revision_response(admitted)
-    refused = admit_experiment_revision_response(refusal)
-
-    assert success == ExperimentRevisionAdmittedResponse(
-        revision_id="sha256:experiment",
-        created=True,
-    )
-    assert refused == RefusalResponse(refusal=refusal)
-
-
-def test_execution_service_language_run_success_contract_is_closed() -> None:
-    payload = {
-        "session_id": "session-1",
-        "revision_id": "sha256:experiment",
-    }
-    artifacts = {"evaluation-run": {"artifact_kind": "evaluation-run"}}
-
-    request = RunExperimentRequest.model_validate(payload)
-    response = RunSuccessResponse(artifacts=artifacts)
-
-    assert request.model_dump(mode="json") == payload
-    assert response.model_dump(mode="json") == {
-        "outcome": "success",
-        "artifacts": artifacts,
-    }
-    with pytest.raises(ValidationError):
-        RunExperimentRequest.model_validate({**payload, "latest": True})
-
-
-def test_execution_service_language_frames_run_results() -> None:
-    artifact = {"artifact_kind": "evaluation-run", "events": [{"value": 7}]}
-    member = PublicationMember(
-        value=artifact,
-        artifact_kind="evaluation-run",
-        wire_schema_identity="sha256:wire-schema",
-        content_identity="sha256:artifact",
-    )
-    refusal = _example_refusal()
-
-    success = run_experiment_revision_response(
-        ExperimentExecutionSuccess(members={"evaluation-run": member})
-    )
-    verdict = run_experiment_revision_response(
-        ExperimentExecutionVerdict(
-            failed_metrics=("damage-floor",),
-            members={"evaluation-run": member},
-        )
-    )
-    refused = run_experiment_revision_response(
-        ExperimentExecutionRefusal(
-            report=refusal,
-            members={"evaluation-run": member},
-        )
-    )
-
-    assert success == RunSuccessResponse(
-        artifacts={"evaluation-run": artifact},
-    )
-    assert verdict == RunVerdictResponse(
-        failed_metrics=["damage-floor"],
-        artifacts={"evaluation-run": artifact},
-    )
-    assert refused == RunRefusalResponse(
-        refusal=refusal,
-        artifacts={"evaluation-run": artifact},
-    )
-    artifact["events"][0]["value"] = 99
-    assert success.artifacts["evaluation-run"]["events"] == [{"value": 7}]
-
-
-def test_execution_service_language_reuses_the_domain_refusal_contract() -> None:
-    refusal = _example_refusal()
-
-    response = RefusalResponse(refusal=refusal)
-    mutated = response.model_dump(mode="json")
-    mutated["refusal"]["transport_extension"] = True
-
-    assert response.refusal is refusal
-    assert "Schema2RefusalReport" in RefusalResponse.model_json_schema()["$defs"]
-    with pytest.raises(ValidationError):
-        RefusalResponse.model_validate(mutated)
-
-
-def test_execution_service_language_run_outcomes_share_one_artifact_shape() -> None:
-    artifacts = {"evaluation-run": {"artifact_kind": "evaluation-run"}}
-    refusal = _example_refusal()
-
-    verdict = RunVerdictResponse(
-        failed_metrics=["damage-per-turn"],
-        artifacts=artifacts,
-    )
-    refused = RunRefusalResponse(refusal=refusal, artifacts=artifacts)
-
-    assert verdict.model_dump(mode="json") == {
-        "outcome": "verdict",
-        "failed_metrics": ["damage-per-turn"],
-        "artifacts": artifacts,
-    }
-    assert refused.model_dump(mode="json") == {
-        "outcome": "refusal",
-        "refusal": refusal.model_dump(mode="json"),
-        "artifacts": artifacts,
-    }
-    assert refused.refusal is refusal
-
-
-def test_execution_service_language_release_session_contract_is_closed() -> None:
-    request = ReleaseExecutionSessionRequest(session_id="session-1")
-    response = ExecutionSessionDeletedResponse(session_id="session-1")
-
-    assert request.model_dump(mode="json") == {"session_id": "session-1"}
-    assert response.model_dump(mode="json") == {
-        "outcome": "success",
-        "session_id": "session-1",
-    }
-    with pytest.raises(ValidationError):
-        ExecutionSessionDeletedResponse.model_validate(
-            {
-                "outcome": "success",
-                "session_id": "session-1",
-                "remaining_revisions": 0,
-            }
-        )
-
-
-@pytest.mark.parametrize(
-    ("code", "message"),
-    [
-        (
-            "unknown_execution_session",
-            "the Execution session does not exist",
-        ),
-        (
-            "unknown_experiment_revision",
-            "the Experiment revision does not exist",
-        ),
-    ],
-)
-def test_execution_service_language_owns_shared_selection_errors(
-    code: ExecutionServiceErrorCode,
-    message: str,
-) -> None:
-    error = execution_service_error(code)
-
-    assert error.model_dump(mode="json") == {
-        "error": {
-            "category": "service",
-            "code": code,
-            "message": message,
-        }
-    }
-
-
-@pytest.mark.parametrize(
-    ("condition", "code"),
-    [
-        (ExecutionSessionNotFound("session-1"), "unknown_execution_session"),
-        (
-            ExperimentRevisionNotFound("sha256:experiment"),
-            "unknown_experiment_revision",
-        ),
-    ],
-)
-def test_execution_service_language_maps_application_selection_conditions(
-    condition: ExecutionSessionNotFound | ExperimentRevisionNotFound,
-    code: ExecutionServiceErrorCode,
-) -> None:
-    error = execution_service_error_from_condition(condition)
-
-    assert error.error.code == code
 
 
 @pytest.fixture(scope="module")
